Arduino Codes Uploaded.
These codes were used to establish a connection between ROS and Arduino. Moreover, some of the Arduino codes are individual meaning that they were used to test the motor control system and to gather the data from wheel encoders. Velocity_Command_Intake_from_ROS.ino is the main file that gets the velocity command from ROS and transforms that into a PWM signal for each of the wheels.